Fix bugs in resave_particles() and pickup_particles() function + Add capability for GADGET4
Fixing bugs in the resave_particles() function and in the function's calling step.

The old version did not load the star information even though the star information was used later.

Fix the order when returning the star information ("svel" should come before "sids")

Also add capability for GADGET4. For the make_pfs_allsnap() function, add the handles that are used by GADGET4 and ARPEO-TNG.